feat: Implement PetList and PetListHeader components for improved pet management UI
- Added PetList component to handle rendering of pet cards with optimized performance.
- Introduced PetListHeader for adding new pets with loading state management.
- Created LoadingOverlay component for consistent loading indicators.
- Refactored PetsScreen to utilize new components and custom hooks for navigation and scrolling.
- Enhanced hooks for better separation of concerns and improved readability.
- Updated index files to export new components and hooks.
