+++ src/components/LoadingOverlay/index.ts
@@ -1,0 +1,1 @@
+export { LoadingOverlay } from './LoadingOverlay';
