+++ src/components/PetList/index.ts
@@ -1,0 +1,1 @@
+export { PetList } from './PetList';
