+++ src/components/PetListHeader/index.ts
@@ -1,0 +1,1 @@
+export { PetListHeader } from './PetListHeader';
