+++ src/components/index.ts
@@ -1,3 +1,6 @@
 export { PetCard } from './PetCard';
 export { default as ScrollIndicator } from './ScrollIndicator';
 export { default as EmptyState } from './EmptyState';
+export { PetListHeader } from './PetListHeader';
+export { LoadingOverlay } from './LoadingOverlay';
+export { PetList } from './PetList';
