+++ src/hooks/index.ts
@@ -1,2 +1,4 @@
-export * from './useNutrition';
-export * from './usePetCalories';
+export { useNutrition } from './useNutrition';
+export { usePetCalories } from './usePetCalories';
+export { usePetScreenNavigation } from './usePetScreenNavigation';
+export { usePetScrolling } from './usePetScrolling';
